Changing all the is_valid_foo queries into is_valid_$LIBRARY_foo in order to achieve name clashes, all their usage in all the libraries I was aware of were updated.
Several small spread changes to make all the classes correct and to
make all the example compilable and (hopefully correctly) running,
with the following exceptions:

- in eiffel-glib: tree_demo does not compile; it was once built as a
  testbed for horrible callbacks hacks; it requires a complete
  redesign. Some other examples show glitches during quit because
  GObject or Glib are not properly shut-down.

- in eiffel-gtk icon_view_demo and paned_examples does not compile. 

Except this "known bugs" I'm committing changes that makes 1206 GTK
classes and 1189 Glib classes correct: I checked it with
check-cluster.sh ; please feel free to improve it and to move to a
more proper place; perhaps $eiffel_libraries/[misc|scripts] could be
meaningful.

I hope this is a step in the right direction for 0.2 release.

  Paolo


git-svn-id: http://svn.gna.org/svn/eiffel-libraries/trunk@702 f7d703df-6b0f-0410-9a7c-ef81f5045915