Various stylistic clean-ups that would help reducing the size of the documentation produced by eiffeldoc.
While trying to produce all the documentation for EWLC eiffeldoc stopped because it encountered a syntax error in the sources. Beside the fact that it took more or less half an hour to process most of EWLC clusters it produced more than 350 Mb of HTML documentation. I'm sure you agree with me that this is not so desiderable, having a 'real' codebase that is ~100 times smaller. Perhaps a custom web-documentation-server a-la-monodoc could be useful.

git-svn-id: http://svn.gna.org/svn/eiffel-libraries/trunk@748 f7d703df-6b0f-0410-9a7c-ef81f5045915