New mock tool.
The start of a new adventure: a real mocker tool, based on the SmartEiffel code to access the complete definition of a class with all its deferred features, including the inherited ones.

Note finished: the tool currently only displays the list of the deferred features it found. The actual mock class need to be generated.