Still working on the new mock tool.
Using the same "signature" idea as the old mock tool as it is an easy way to gather information using a VISITOR; but using the standard SmartEiffel coding style (try to lose as little memory as possible).

The MOCK class is now completely generated; the EXPECT class is still to do.