Make the GC dispose call at exit optional
We are at the frontier between TDD and DbC.

The dispose method is not meant to perform significant computations.
In the Liberty Eiffel standard library it is only used as an extra check at the exit of the process.
It means that it is useless in boost mode.
It also means that mock objects may need some extra checks after the test itself is finished.

So the dispose call at exit is now only called if non-boost modes.

Furthermore, the mock framework also forbids that extra call.