change routine_then handling
Some frantic debugging and cleanup, and the conclusion that routine_then must disappear as soon as possible.
Not too early, since documentation tools must be able to reconstitute the original text.
The best position is in specialize_and_check, when the compiler has its first chance to modify the AST.