add some query options to ace_check
This might be better as a single query option in the future, or perhaps
a separate config tool.

At some time in the distant past ace_check used to output a prettified
version of the file with expanded clusters.  I've removed the disabled
code as this functionality would be better provided by a combination of
the pretty command and the aforementioned expanded query option.